# docs/swagger.yaml
basePath: /notifications/api
definitions:
  CoreUserRef:
    properties:
      name:
        type: string
      user_id:
        type: string
    type: object
...
  Recipient:
    properties:
      name:
        type: string
      notification_disabled:
        type: boolean
      user_id:
        type: string
    type: object
...
  Topic:
    properties:
      app_id:
        type: string
      date_created:
        type: string
      date_updated:
        type: string
      description:
        type: string
      name:
        type: string
      org_id:
        type: string
    type: object
...
  model.Message:
    description: wraps all needed information for the notification
    properties:
      body:
        type: string
      data:
        additionalProperties:
          type: string
        type: object
      date_created:
        type: string
      date_updated:
        type: string
      id:
        type: string
      priority:
        type: integer
      recipients:
        items:
          $ref: '#/definitions/Recipient'
        type: array
      recipients_criteria_list:
        items:
          $ref: '#/definitions/model.RecipientCriteria'
        type: array
      sender:
        $ref: '#/definitions/model.Sender'
      subject:
        type: string
      topic:
        type: string
    type: object
  model.RecipientCriteria:
    properties:
      app_platform:
        type: string
      app_version:
        type: string
    type: object
  model.Sender:
    properties:
      type:
        description: user or system
        type: string
      user:
        $ref: '#/definitions/CoreUserRef'
    type: object
...
host: localhost
info:
  contact: {}
  description: Rokwire Notifications Building Block API Documentation.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Rokwire Notifications Building Block API
  version: 1.1.8
paths:
  /admin/app_platforms:
    get:
      description: Gets all available app platforms
      operationId: GetAllAppPlatforms
      produces:
      - text/plain
      responses:
        "200":
          description: ""
      security:
      - AdminUserAuth: []
      tags:
      - Admin
  /admin/app_versions:
    get:
      description: Gets all available app versions
      operationId: GetAllAppVersions
      produces:
      - text/plain
      responses:
        "200":
          description: ""
      security:
      - AdminUserAuth: []
      tags:
      - Admin
  /admin/message:
    post:
      consumes:
      - application/json
      description: Creates a message
      operationId: CreateMessage
      parameters:
      - description: body json
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.Message'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Message'
      security:
      - AdminUserAuth: []
      tags:
      - Admin
    put:
      consumes:
      - application/json
      description: Updates a message
      operationId: UpdateMessage
      parameters:
      - description: body json
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.Message'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Message'
      security:
      - AdminUserAuth: []
      tags:
      - Admin
  /admin/message/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a message with id
      operationId: DeleteMessage
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: ""
      security:
      - AdminUserAuth: []
      tags:
      - Admin
    get:
      consumes:
      - application/json
      description: Retrieves a message by id
      operationId: GetMessage
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Message'
      security:
      - AdminUserAuth: []
      tags:
      - Admin
  /admin/messages:
    get:
      description: Gets all messages
      operationId: GetMessages
      parameters:
      - description: user - filter by user
        in: query
        name: user
        type: string
      - description: topic - filter by topic
        in: query
        name: topic
        type: string
      - description: offset
        in: query
        name: offset
        type: string
      - description: limit - limit the result
        in: query
        name: limit
        type: string
      - description: 'order - Possible values: asc, desc. Default: desc'
        in: query
        name: order
        type: string
      - description: start_date - Start date filter in milliseconds as an integer
          epoch value
        in: query
        name: start_date
        type: string
      - description: end_date - End date filter in milliseconds as an integer epoch
          value
        in: query
        name: end_date
        type: string
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Message'
            type: array
      security:
      - AdminUserAuth: []
      tags:
      - Admin
  /admin/topic:
    put:
      description: Updated the topic.
      operationId: UpdateTopic
      parameters:
      - description: body json
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/Topic'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Topic'
      security:
      - AdminUserAuth: []
      tags:
      - Admin
  /admin/topics:
    get:
      description: Gets all topics
      operationId: AdminGetTopics
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/Topic'
            type: array
      security:
      - AdminUserAuth: []
      tags:
      - Admin
...
  /message:
    post:
      consumes:
      - application/json
      description: Creates a message. Message without subject and body will be interpreted
        as a data massage and it won't be stored in the database
      operationId: createMessage
      parameters:
      - description: body json
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.Message'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Message'
      security:
      - UserAuth: []
      tags:
      - Client
...
  /topics:
    get:
      description: Gets all topics
      operationId: GetTopics
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/Topic'
            type: array
      security:
      - RokwireAuth: []
      tags:
      - Client
...
schemes:
- https
securityDefinitions:
  AdminUserAuth:
    in: header (add admin core access token with Bearer prefix to the Authorization
      value. The token must contain notifications_admin as a permission)
    name: Authorization
    type: apiKey
  InternalAuth:
    in: header
    name: INTERNAL-API-KEY
    type: apiKey
  RokwireAuth:
    in: header
    name: ROKWIRE-API-KEY
    type: apiKey
  UserAuth:
    in: header (add core access token with Bearer prefix to the Authorization value.
      The token must represent either anonymous or authenticated user )
    name: Authorization
    type: apiKey
swagger: "2.0"